fixed bug #407192 with cue not splitting the last file

+++ mp3splt-project/trunk/mp3splt-gtk/src/tree_tab.h
@@ -99,6 +99,6 @@ void handle_detached_event (GtkHandleBox *handlebox,
                             gpointer data);
 GtkWidget *create_choose_splitpoints_frame(GtkTreeView *tree_view);
 void garray_to_array(GArray *spltpoints, 
-                     gint *hundredth);
+                     glong *hundredth);
 void put_splitpoints_in_the_state(splt_state *state);
 GtkTreeView *create_tree_view();
